change directory name
