+++ click_stream/__init__.py
@@ -1,0 +1,1 @@
+import click_stream
