--- clickstream/constants.py
@@ -1,24 +1,0 @@
-WEEKDAY_DICT = {
-    0: 'MON',
-    1: 'TUE',
-    2: 'WED',
-    3: 'THU',
-    4: 'FRI',
-    5: 'SAT',
-    6: 'SUN',
-}
-
-MONTH_DICT = {
-    1:'Jan', 
-    2:'Feb', 
-    3:'Mar', 
-    4:'Apr', 
-    5:'May', 
-    6:'Jun', 
-    7:'Jul',
-    8:'Aug', 
-    9:'Sep', 
-    10:'Oct', 
-    11:'Nov', 
-    12:'Dec'
-}
