+++ query_string/__init__.py
@@ -1,0 +1,1 @@
+import query_string
