+++ query_string/query_string.md
@@ -1,0 +1,1 @@
+# Qurry String 전처리 방법
