--- querystring/query_string.md
@@ -1,1 +1,0 @@
-# Qurry String 전처리 방법
